##### 0908 - @Validated 要放在class上而不是方法上 - @Positive(message = "{id.positive}") 在resources 的 ValidationMessages.properties 去配置 - 通过categoryId 查询

+++ src/main/java/com/xl/missyou/model/Spu.java
@@ -17,8 +17,8 @@ public class Spu extends BaseEntity {
     private Long id;
     private String title;
     private String subtitle;
-    private int categoryId;
-    private Integer rootCategoryId;
+    private Long categoryId;
+    private Long rootCategoryId;
     private Boolean online;
     private String price;
     private Integer sketchSpecId;
